post creation: put in user ID for authorship
This is necessary because we do not have the 'Author' field in the form
for new post creation; this means that the application will not receive
the 'Author' parameter when we click on "Create Post".

Also, we remove the 'author: current_user.id' value for the @post
instance variable @post, because we do not need to have that variable at
all in the form before we click on "Create Post".