correct_user: use unique definition across controllers
This way, we define what a "correct" user is, depending on what model we
are concerned with. If this kind of definition becomes too repetitive,
we could probably put it back into the sessions controller and pass the
context into it, so that we have a big case/when/else statement to take
care of the different model contexts. However, for our purposes,
redefining it per model is sufficient.